# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EasySell Documentation",
        "description": "EasySell description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "EasySell API Server not authorized"
        }
    ],
    "paths": {
        "/products/nomenclatures/toArchive": {
            "post": {
                "tags": [
                    "Archives"
                ],
                "summary": "Nomenclatures to Archive",
                "operationId": "40e3484c2f66b741ed32891c0730e082",
                "requestBody": {
                    "description": "Nomenclatures to Archive",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nomenclatures": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "collectionFormat": "multi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}